requirements.txt
streamlit
streamlit-gsheets-connection
pandas
openai